% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appendmodel_function.R
\name{appendmodel}
\alias{appendmodel}
\alias{appendmodel.formula}
\title{Change left-hand side of formula}
\usage{
appendmodel(object, prepend = "", append = "")
}
\arguments{
\item{object}{object containing formula of interest}

\item{prepend}{character that will precede the model expression}

\item{append}{character that will succede the model expression}
}
\description{
Prepends or appends to the right-hand-side of a formula
}
\examples{
appendmodel(~ a + b, "exp(", ")")
appendmodel(z ~ a + b, " y + ")
}
\author{
Kim Rand
}
