% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_target_function.R
\name{change_target}
\alias{change_target}
\alias{change_target.formula}
\title{Change left-hand side of formula}
\usage{
change_target(object, newtarget, oldtarget = NULL)
}
\arguments{
\item{object}{object containing formula of interest}

\item{newtarget}{character for new left-hand side}

\item{oldtarget}{character to restrict to specific existing target. If NULL, any target will be changed.}
}
\description{
Changes the left-hand side of a formula
}
\examples{
formula1 <- oldtarget ~ x + y
change_target(formula1, "newtarget")
change_target(formula1, "newtarget", "not_this_target")
}
\author{
Kim Rand-Hendriksen
}
