% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cont_normal_function.R
\name{cont_normal}
\alias{berkson_w,}
\alias{berkson_w_aggr,}
\alias{cont_logistic,}
\alias{cont_normal}
\alias{dich_logistic,}
\alias{dich_normal,}
\alias{mse_p}
\title{Likelihood functions for xregControl}
\usage{
cont_normal(d_df, sigma_est = 0)
}
\arguments{
\item{d_df}{data.frame provided by xreg.}

\item{sigma_est}{Subsequent parameters define required formulas to genrete necessary variables, such as standard deviations. If these are not manually provided (which could be useful to model heteroscedasticity, etc.), xreg will get them from the functions.}
}
\value{
\code{cont_normal} normal-distribution-based likelihood for continuous variables

\code{cont_logistic} logistic-distribution-based likelihood for continuous variables

\code{dich_normal} normal-distribution-based likelihood for dichotomous variables (probit)

\code{dich_logistic} logistic-distribution-based likelihood for dichotomous variables (logit)

\code{berkson_w} chi_square distance

\code{berkson_w_aggr} weighted aggregation for chi-square

\code{mse_p} means square error
}
\description{
Likelihood functions to use with xregControl
}
\details{
Likelihood functions for data.frame
}

