% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglik_functions.R
\name{cont_ols}
\alias{cont_ols}
\title{OLS-minimizer, used by xregControl and xreg}
\usage{
cont_ols(d_df)
}
\arguments{
\item{d_df}{data.frame provided by xreg.}
}
\description{
Likelihood function to use with xregControl
}
\details{
Likelihood functions for data.frame
}
