% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cont_r_normal_function.R
\name{cont_r_normal}
\alias{cont_r_normal}
\title{Normal-distribution-based random intercept likelihood function for xregControl and xreg}
\usage{
cont_r_normal(d_df, sigma_est = 0, omega_est = 0, internal_id = 0)
}
\arguments{
\item{d_df}{data.frame provided by xreg.}

\item{sigma_est}{Within-unit variance parameter, may be specified as a function to vary between observations. Defaults to exp(LN_SIGMA)}

\item{omega_est}{Between-unit variance parameter, must be a fixed result for all observations, as only the first instance will be used. Defaults to exp(LN_OMEGA)}
}
\description{
Likelihood function to use with xregControl
}
\details{
Likelihood functions for data.frame
}
