% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglik_functions.R
\name{dich_logistic}
\alias{dich_logistic}
\title{Logistic-distribution-based likelihood function for dichotomous data, used by xregControl and xreg}
\usage{
dich_logistic(d_df, theta_est = 0)
}
\arguments{
\item{d_df}{data.frame provided by xreg.}

\item{theta_est}{Subsequent parameters define required formulas to genrete necessary variables, such as standard deviations. If these are not manually provided (which could be useful to model heteroscedasticity, etc.), xreg will get them from the functions.}
}
\description{
Likelihood function to use with xregControl
}
\details{
Likelihood functions for data.frame
}
