% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglik_functions.R
\name{dich_loglik}
\alias{dich_loglik}
\title{Probability likelihood over probabilities, used by xregControl and xreg}
\usage{
dich_loglik(d_df)
}
\arguments{
\item{d_df}{data.frame provided by xreg.}
}
\description{
Likelihood function to use with xregControl
}
\details{
Likelihood functions for data.frame
}
