% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglik_functions.R
\name{dich_r_logistic}
\alias{dich_r_logistic}
\title{Logistic-distribution-based likelihood function for dichotomous data with random intercepts, used by xregControl and xreg}
\usage{
dich_r_logistic(d_df, theta_est = 0, omega_est = 0, internal_id = 0)
}
\arguments{
\item{d_df}{data.frame provided by xreg.}

\item{theta_est}{Subsequent parameters define required formulas to genrete necessary variables, such as standard deviations. If these are not manually provided (which could be useful to model heteroscedasticity, etc.), xreg will get them from the functions.}

\item{omega_est}{Between-unit variance parameter, must be a fixed result for all observations, as only the first instance will be used. Defaults to exp(LN_OMEGA)}

\item{sigma_est}{Within-unit variance parameter, may be specified as a function to vary between observations. Defaults to exp(LN_SIGMA)}
}
\description{
Likelihood function to use with xregControl
}
\details{
Likelihood functions for data.frame
}
