% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyreg_function.R
\name{hyreg}
\alias{hyreg}
\title{Flexible maximum-likelihood-based regression function}
\usage{
hyreg(formula, df, datatype = "method", init = numeric(),
  contdist = "normal", dichdist = "logistic", hetcont = NULL,
  hetdich = NULL, ul = Inf, ll = -Inf, lntheta = T, lnsigma = T,
  dichformula = NULL, ...)
}
\arguments{
\item{formula}{a formula or list of formulas to be passed to xregControl and xreg. If a single formula is provided, the left side argument must match the dependent variable. If interval regression is intended, the upper and lower bounds should be assigned to two columns starting with a common name, e.g. value.lb and value.ub. For point estimates, these should be identical.}

\item{df}{data.frame containing both the continuous and dichotomous data. Will be split into two data.frames using a column matching the name provided in the datatype argument.}

\item{datatype}{character defining the column name separating dichotomous from continuous data. Continuous data should be interpretable as TRUE in as.logical, and dichotomous as FALSE}

\item{init}{Optional named list or vector of numeric start values, to override those provided in the xregControl obects. Previously estimated xreg-objects can be used, in which case fitted values will be extracted.}

\item{contdist}{character option to determine if the continuous model should use normal or logistic distribution. Options are "normal", and "logistic". "normal" instructs xregControl to provide cont_normal(), logistic provides cont_logistic(). Default is "normal".}

\item{dichdist}{character option to determine if the dichotomous model should use normal or logistic distribution. Options are "normal", and "logistic". "normal" instructs xregControl to provide dich_normal(), logistic provides dich_logistic(). Default is "logistic".}

\item{hetcont}{optional formula for fitting hetreoscedastic standard deviations for the continuous model. Parameters sharing names with the ones provided in the formula argument will be the same. If this is not intended, different parmaeter names are required.}

\item{hetdich}{optional formula for fitting hetreoscedastic standard deviations for the dichotomous model. Requires dichotomous distributionn to be "normal". Parameters sharing names with the ones provided in the formula argument will be the same. If this is not intended, different parmaeter names are required.}

\item{ul}{optional upper bound value for the dependent variable. Values over ul will be set to an interval between the observed and Inf, making the fitting similar to tobit regression. Defaults to Inf.}

\item{ll}{optional lower bound value for the dependent variable. Values below ll will be set to an interval between the observed and -Inf, making the fitting similar to tobit regression. Defaults to -Inf.}

\item{lntheta}{If TRUE, parameters or formulas for theta will be exponentiated using exp() prior to fitting, to match the default in the STATA hyreg command. Defaults to TRUE.}

\item{lnsigma}{If TRUE, parameters or formulas for sigma will be exponentiated using exp() prior to fitting, to match the default in the STATA hyreg command. Defaults to TRUE.}

\item{dichformula}{optional formula argument for the dichotomous dataset. If not provided, the formula for dichotomous data will be the same as for continuous.}

\item{...}{optional arguments to be forwarded to xregControl, xreg, x_mle, or optim}
}
\description{
Wrapper for xregControl and xreg to match the functionality of the STATA hyreg function.
}
\examples{
control_continuous <- xregControl(formulas = list(x ~ y * YVAR + z + ZVAR, value ~ INTERCEPT + x * XVAR), start_values = c(INTERCEPT = 0, XVAR = 1, YVAR = 1, ZVAR = 1), p_fun = cont_normal, name = "CONTINUOUS")
control_dichotomos <- xregControl(formulas = list(value ~ INTERCEPT2 + z * ZVAR), start_values = c(INTERCEPT2 = 0, ZVAR = 1), p_fun = dich_logistic, name = "DICHOTOMOUS")
joint_control <- c(control_continuous, control_dichotomous)
xreg(controlList = joint_control, dataList = list(CONTINUOUS = df1, DICHOTOMUS = df2))
}
\author{
Kim Rand-Hendriksen
}
