% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.xreg_function.R
\name{predict.xreg}
\alias{predict.xreg}
\title{Predict function for xreg}
\usage{
\method{predict}{xreg}(object, newdata = NULL, type = c("link", "response",
  "terms"), return_lik_df = FALSE, return_vector = FALSE)
}
\arguments{
\item{object}{xreg-object}

\item{newdata}{Data.frame with new data, or list of data.frames with new data}

\item{type}{determines what type of predictions are wanted. Not yet implemented}

\item{return_lik_df}{logical argument, if true, the full data.frame with likelihood-estimates will be returned}

\item{return_vector}{logical argument, if true, a single vector of predicted values will be returned}
}
\description{
Predict function for xreg-objects.
}
\author{
Kim Rand-Hendriksen
}

