% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.x_mle_function.R
\name{print.x_mle}
\alias{print.x_mle}
\alias{summary.x_mle}
\title{Print x_mle-object}
\usage{
\method{print}{x_mle}(x_mle)
}
\arguments{
\item{x_mle}{x_mle object}
}
\description{
Prints important results from an x_mle-object
}
\author{
Kim Rand-Hendriksen
}

