% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/try0_function.R
\name{try0}
\alias{try0}
\title{Wrapper for try which returns second argument if not succesful}
\usage{
try0(expr, ifError = NULL)
}
\arguments{
\item{expr}{expression to be tested}

\item{ifError}{return value if test fails}
}
\description{
Returns second argument if expression in first argument fails.
}
\author{
Kim Rand-Hendriksen
}

