% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x_mle_function.R
\name{x_mle}
\alias{x_mle}
\title{Modification of the mle function from the stats4 package}
\usage{
x_mle(minuslogl, start = formals(minuslogl), method = "BFGS",
  fixed = list(), nobs, solve_hessian = T, return_first = FALSE,
  lower = -Inf, upper = Inf, ...)
}
\arguments{
\item{minuslogl}{Function to be minimized}

\item{start}{list or named vector of start values for parametres to be fitted}

\item{method}{Method to be used by \code{\link[stats]{optim}}.}

\item{fixed}{list of fixed named parameters (will not be passed to \code{\link[stats]{optim}} for fitting).}

\item{nobs}{number of observations. Not used presently}

\item{solve_hessian}{If false, will not try to calculate standard errors. Default is TRUE}

\item{return_first}{If set to TRUE, will return result of minusloglik using values supplied to start}

\item{lower}{named list of lower bounds for fitting of values. Requires method L-BFGS-B}

\item{upper}{named list of upper bounds for fitting of values. Requires method L_BFGS-B}

\item{...}{Arguments to be passed on to minuslogl or optim}
}
\description{
Modified version of the mle function from stats4. Allows arbitrary parameters in function provided to parmaeter minuslogl. Intended for internal use by xreg.
}
\author{
Original funciton by R Core Team and contributors worldwide. Edits to work properly with xreg by Kim Rand-Hendriksen
}
