% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlik_function.R
\name{xlik}
\alias{xlik}
\alias{summary.xreg,}
\alias{print.xreg}
\title{Internal function used by xreg}
\usage{
xlik(
  ...,
  controlList,
  dataList,
  print_sum = FALSE,
  return_lik_df = FALSE,
  return_first_df = FALSE,
  latent_classes = 0,
  latent_class_parameters = character()
)
}
\arguments{
\item{...}{optional arguments to be forwarded to x_mle, optim, xregControl, etc.}

\item{controlList}{one or more xregControl-objects (concatenated using c() ) to control maximum-likelihood estimation for different types or sets of data provided in dataList. Can alternatively provide a single formula or list of formulas, in which case an xregControl object will be generated on the fly. An xreg-object can be provided, in which case the xregControlList will be extracted. See examples}

\item{dataList}{Named list of data.frames for which maximum sum-likelihood should be estimated. Names must correspond to names of xregControl objects. If no names are provided, order will be used to decide which xregControl-object to use.}

\item{print_sum}{Optional argument to force continuous printing of the progress of fitting. Can be useful for debugging.}

\item{return_lik_df}{if true, will return the data.frame(s) on which likelihood are estimated}

\item{return_first_df}{if true, eill return the first fit to initial values}

\item{latent_classes}{Optional integer parameter defining the number of latent classes to be fitted}

\item{latent_class_parameter}{Optional character vector with names corresponding to the parameters that are to be varied in the latent classes (if latent_classes > 1)}
}
\description{
Allows a list of formulas and fitting functions to be used for each type or set of data.
}
\author{
Kim Rand
}
